Added new database connection - Bento-Auth - Bento-User
